Starting to implement the API
